overhaul/speed up ucc test

+++ inbm/integration-reloaded/test/ucc/INSTALL_FRAMEWORK_ucc.sh
@@ -3,13 +3,8 @@
 set -euxo pipefail
 
 cd /vagrant/input
-cp /scripts/succeed_rpm_cert.pem .
 
-# Update apt database
 apt-get update
-
-# Remove/uninstall prereqs to ensure installer deals with them
-apt-get purge -y --allow-change-held-packages lxc mosquitto cryptsetup less docker-compose
 
 # Simulate user calling the installer
 mkdir "install TC" # test install dir with spaces
@@ -20,21 +15,30 @@ rm -rf /etc/intel-manageability/public/cloudadapter-agent
 mkdir -p /etc/intel-manageability/public/cloudadapter-agent
 apt-get purge -y docker-ce docker-ce-cli || true
 rm -rf /var/lib/apt/lists/*
-sudo -H UCC_MODE=true ACCEPT_INTEL_LICENSE=true bash -x ./install-tc.sh
 
-date
+## No TPM simulator in quicker mode
+## DEV_MODE=true INSTALL_TPM2_SIMULATOR=false ACCEPT_INTEL_LICENSE=true bash -x ./install-tc.sh
+## date
+
+# Update shell to force dpkg to use bash during installation.
+echo "dash dash/sh boolean false" | debconf-set-selections
+if ! dpkg-reconfigure dash -f noninteractive; then
+  echo "Unable to configure environment (dash->bash)"
+  exit 1
+fi
+
+
+tar -zxvf *.preview.tar.gz
+
+sudo -H UCC_MODE=true DEV_MODE=true INSTALL_TPM2_SIMULATOR=false ACCEPT_INTEL_LICENSE=true bash -x ./install-tc.sh
 
 for i in cloudadapter ; do
   sed -i 's/ERROR/DEBUG/g' /etc/intel-manageability/public/"$i"-agent/logging.ini
 done
 
-
-cp /scripts/iotg_inb_developer.conf /etc/intel_manageability.conf
 cp /scripts/inb_fw_tool_info.conf /etc/firmware_tool_info.conf
-touch /etc/intel-manageability/public/cloudadapter-agent/iot-dispatcher.cfg
 
-sudo -H NO_CLOUD=x PROVISION_TPM=disable NO_OTA_CERT=1 LOCAL_MQTT_PORT=9999 bash -x /usr/bin/provision-tc
-
+NO_CLOUD=1 PROVISION_TPM=auto NO_OTA_CERT=1 TELIT_HOST="localhost" bash -x /usr/bin/provision-tc
 # NOTE: this has to be redone if we change the template or the
 # inb-provision-cloud binary. Alternately we could create a script
 # interface to inb-provision-cloud.
@@ -81,30 +85,9 @@ sudo dd of=/etc/intel-manageability/secret/cloudadapter-agent/adapter.cfg <<EOF
 }
 EOF
 
-
-touch /etc/intel-manageability/public/cloudadapter-agent/iot-dispatcher.cfg
-
 sleep 5
 echo All processes:
 ps -ax
 echo Cloudadapter:
 ps -ax | grep cloudadapter
 
-for i in /etc/intel-manageability/secret/* ; do
-    BASENAME="$(basename $i)"
-    if ! [ "$BASENAME" == "lost+found" ] ; then
-        : Check permissions of "$i"
-        stat -c %A "$i" | egrep -- "^drwxr-x---$"
-
-        : Check group of "$i"
-        stat -c %G "$i" | egrep -- "^$BASENAME$"
-
-        : Check owner of "$i"
-        stat -c %U "$i" | egrep -- "^root$"
-    fi
-done
-
-systemctl disable mqtt-keygen || true
-rm -rf /lib/systemd/system/mqtt-keygen.service
-
-systemctl start mqtt inbm-cloudadapter
